+++ packages/core/src/constants/chainId.ts
@@ -4,7 +4,10 @@ export enum ChainId {
   Rinkeby = 4,
   Goerli = 5,
   Kovan = 42,
+  BSC = 56,
   xDai = 100,
+  Polygon = 137,
+  Mumbai = 80001,
   Localhost = 1337,
   Hardhat = 31337,
 }
@@ -15,7 +18,10 @@ export const CHAIN_NAMES = {
   [ChainId.Kovan]: 'Kovan',
   [ChainId.Rinkeby]: 'Rinkeby',
   [ChainId.Goerli]: 'Goerli',
+  [ChainId.BSC]: 'BSC',
   [ChainId.xDai]: 'xDai',
+  [ChainId.Polygon]: 'Polygon',
+  [ChainId.Mumbai]: 'Mumbai',
   [ChainId.Localhost]: 'Localhost',
   [ChainId.Hardhat]: 'Hardhat',
 }
@@ -26,7 +32,10 @@ export const MULTICALL_ADDRESSES = {
   [ChainId.Rinkeby]: '0x42ad527de7d4e9d9d011ac45b31d8551f8fe9821',
   [ChainId.Goerli]: '0x77dca2c955b15e9de4dbbcf1246b4b85b651e50e',
   [ChainId.Kovan]: '0x2cc8688c5f75e365aaeeb4ea8d6a480405a48d2a',
+  [ChainId.BSC]: '0x41263cba59eb80dc200f3e2544eda4ed6a90e76c',
   [ChainId.xDai]: '0xb5b692a88bdfc81ca69dcb1d924f59f0413a602a',
+  [ChainId.Polygon]: '0x11ce4B23bD875D7F5C6a31084f55fDe1e9A87507',
+  [ChainId.Mumbai]: '0x08411ADd0b5AA8ee47563b146743C13b3556c9Cc',
 }
 
 export const TEST_CHAINS = [
@@ -34,6 +43,7 @@ export const TEST_CHAINS = [
   ChainId.Kovan,
   ChainId.Rinkeby,
   ChainId.Goerli,
+  ChainId.Mumbai,
   ChainId.Localhost,
   ChainId.Hardhat,
 ]