+++ packages/core/src/helpers/chain.ts
@@ -20,8 +20,14 @@ export function getExplorerAddressLink(address: string, chainId: ChainId) {
     case ChainId.Rinkeby:
     case ChainId.Goerli:
       return `https://${etherscanNetworkPrefix(chainId)}etherscan.io/address/${address}`
+    case ChainId.BSC:
+      return `https://bscscan.com/address/${address}`
     case ChainId.xDai:
       return `https://blockscout.com/poa/xdai/address/${address}/transactions`
+    case ChainId.Polygon:
+      return `https://explorer-mainnet.maticvigil.com/address/${address}/transactions`
+    case ChainId.Mumbai:
+      return `https://explorer-mumbai.maticvigil.com/address/${address}/transactions`
   }
 }
 
@@ -33,8 +39,14 @@ export function getExplorerTransactionLink(transactionHash: string, chainId: Cha
     case ChainId.Rinkeby:
     case ChainId.Goerli:
       return `https://${etherscanNetworkPrefix(chainId)}etherscan.io/tx/${transactionHash}`
+    case ChainId.BSC:
+      return `https://bscscan.com/tx/${transactionHash}`
     case ChainId.xDai:
       return `https://blockscout.com/poa/xdai/tx/${transactionHash}/internal-transactions`
+    case ChainId.Polygon:
+      return `https://explorer-mainnet.maticvigil.com/tx/${transactionHash}/internal-transactions`
+    case ChainId.Mumbai:
+      return `https://explorer-mumbai.maticvigil.com/tx/${transactionHash}/internal-transactions`
   }
 }