+++ packages/core/src/model/TransactionStatus.ts
@@ -1,0 +1,16 @@
+import { ChainId } from '../constants'
+import { TransactionResponse, TransactionReceipt } from '@ethersproject/abstract-provider'
+
+export type TransactionState = 'None' | 'Mining' | 'Success' | 'Fail' | 'Exception'
+
+export interface TransactionStatus {
+  status: TransactionState
+  transaction?: TransactionResponse
+  receipt?: TransactionReceipt
+  chainId?: ChainId
+  errorMessage?: string
+}
+
+export function transactionErrored(transaction: TransactionStatus) {
+  return 'errorMessage' in transaction
+}