+++ packages/core/src/model/index.ts
@@ -1,3 +1,5 @@
 export * from './Currency'
 export * from './CurrencyValue'
 export * from './config/Config'
+export * from './TransactionStatus'
+export * from './TransactionOptions'