+++ packages/core/src/providers/chainState/provider.tsx
@@ -21,20 +21,6 @@ export function ChainStateProvider({ children, multicallAddresses }: Props) {
   const [calls, dispatchCalls] = useReducer(callsReducer, [])
   const [state, dispatchState] = useReducer(chainStateReducer, {})
 
-  const addCalls = useCallback(
-    (calls: ChainCall[]) => {
-      dispatchCalls({ type: 'ADD_CALLS', calls })
-    },
-    [dispatchCalls]
-  )
-
-  const removeCalls = useCallback(
-    (calls: ChainCall[]) => {
-      dispatchCalls({ type: 'REMOVE_CALLS', calls })
-    },
-    [dispatchCalls]
-  )
-
   const [debouncedCalls, debouncedId] = useDebouncePair(calls, chainId, 50)
   const uniqueCalls = debouncedId === chainId ? getUnique(debouncedCalls) : []
   // used for deep equality in hook dependencies
@@ -80,7 +66,7 @@ export function ChainStateProvider({ children, multicallAddresses }: Props) {
   }, [library, blockNumber, chainId, multicallAddress, uniqueCallsJSON])
 
   const value = chainId !== undefined ? state[chainId] : undefined
-  const provided = { value, multicallAddress, addCalls, removeCalls }
+  const provided = { value, multicallAddress, dispatchCalls }
 
   return <ChainStateContext.Provider value={provided} children={children} />
 }