+++ packages/example/src/App.tsx
@@ -5,9 +5,11 @@ import { Page } from './components/base/base'
 import { TopBar } from './components/TopBar'
 import { GlobalStyle } from './global/GlobalStyle'
 import { Balance } from './pages/Balance'
+import { Prices } from './pages/Prices'
 import { Block } from './pages/Block'
 import { Tokens } from './pages/Tokens'
 import { Transactions } from './pages/Transactions'
+import { SendEtherPage } from './pages/SendEtherPage'
 
 export function App() {
   return (
@@ -17,8 +19,10 @@ export function App() {
         <TopBar />
         <Switch>
           <Route exact path="/balance" component={Balance} />
+          <Route exact path="/prices" component={Prices} />
           <Route exact path="/block" component={Block} />
           <Route exact path="/tokens" component={Tokens} />
+          <Route exact path="/send" component={SendEtherPage} />
           <Route exact path="/transactions" component={Transactions} />
           <Redirect exact from="/" to="/balance" />
         </Switch>