+++ packages/example/src/pages/Transactions.tsx
@@ -1,30 +1,27 @@
 import React from 'react'
-import { useEthers } from '@usedapp/core'
 import styled from 'styled-components'
 import { Container, MainContent, Section, SectionRow } from '../components/base/base'
-import { Button } from '../components/base/Button'
-import { DepositEth, WithdrawEth } from '../components/Notifications/Forms'
-import { NotificationsList, TransactionsList } from '../components/Notifications/History'
+
+import { DepositEth, WithdrawEth } from '../components/Transactions/Forms'
+import { NotificationsList, TransactionsList } from '../components/Transactions/History'
 import { Title } from '../typography/Title'
 
-export function Transactions() {
-  const { activateBrowserWallet, deactivate, account } = useEthers()
+import { AccountButton } from '../components/account/AccountButton'
 
+export function Transactions() {
   return (
     <MainContent>
       <Container>
         <Section>
           <SectionRow>
             <Title>Transactions</Title>
-            {account && <Button onClick={deactivate}>Disconnect</Button>}
-            {!account && <Button onClick={activateBrowserWallet}>Connect</Button>}
+            <AccountButton />
           </SectionRow>
           <TableGrid>
             <DepositEth />
             <WithdrawEth />
-            <TransactionsList />
-            <NotificationsList />
           </TableGrid>
+          <NotificationsList />
         </Section>
       </Container>
     </MainContent>